requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
networkx==3.3
pandas==2.2.2
python-dotenv==1.0.1
openai==1.35.3
google-generativeai==0.7.1
pydantic==2.8.2
